backend/requirements.txt
Flask==3.1.0
Flask-CORS==5.0.0
openai==1.54.5
langchain==0.3.7
langchain-openai==0.2.9
requests==2.32.3
python-dotenv==0.21.0
pprintpp==0.4.0